updating pac4j to 5.x

+++ legend-sdlc-server/src/test/java/org/finos/legend/sdlc/server/gitlab/api/GitLabWorkspaceApiTestResource.java
@@ -216,8 +216,8 @@ public class GitLabWorkspaceApiTestResource
 
         GitLabProjectId sdlcGitLabProjectId = GitLabProjectId.parseProjectId(projectId);
         MergeRequestApi mergeRequestApi = gitLabMemberUserContext.getGitLabAPI().getMergeRequestApi();
-        Integer parsedMergeRequestId = Integer.parseInt(reviewId);
-        Integer gitlabProjectId = sdlcGitLabProjectId.getGitLabId();
+        Long parsedMergeRequestId = Long.parseLong(reviewId);
+        Long gitlabProjectId = sdlcGitLabProjectId.getGitLabId();
 
         String requiredStatus = "can_be_merged";
         CallUntil<MergeRequest, GitLabApiException> callUntil = CallUntil.callUntil(
@@ -353,8 +353,8 @@ public class GitLabWorkspaceApiTestResource
 
         GitLabProjectId sdlcGitLabProjectId = GitLabProjectId.parseProjectId(projectId);
         MergeRequestApi mergeRequestApi = gitLabMemberUserContext.getGitLabAPI().getMergeRequestApi();
-        Integer parsedMergeRequestId = Integer.parseInt(reviewId);
-        Integer gitlabProjectId = sdlcGitLabProjectId.getGitLabId();
+        Long parsedMergeRequestId = Long.parseLong(reviewId);
+        Long gitlabProjectId = sdlcGitLabProjectId.getGitLabId();
 
         String requiredStatus = "can_be_merged";
         CallUntil<MergeRequest, GitLabApiException> callUntil = CallUntil.callUntil(
@@ -577,8 +577,8 @@ public class GitLabWorkspaceApiTestResource
 
         GitLabProjectId sdlcGitLabProjectId = GitLabProjectId.parseProjectId(projectId);
         MergeRequestApi mergeRequestApi = gitLabMemberUserContext.getGitLabAPI().getMergeRequestApi();
-        Integer parsedMergeRequestId = Integer.parseInt(reviewId);
-        Integer gitlabProjectId = sdlcGitLabProjectId.getGitLabId();
+        Long parsedMergeRequestId = Long.parseLong(reviewId);
+        Long gitlabProjectId = sdlcGitLabProjectId.getGitLabId();
 
         String requiredStatus = "can_be_merged";
         CallUntil<MergeRequest, GitLabApiException> callUntil = CallUntil.callUntil(
@@ -718,8 +718,8 @@ public class GitLabWorkspaceApiTestResource
 
         GitLabProjectId sdlcGitLabProjectId = GitLabProjectId.parseProjectId(projectId);
         MergeRequestApi mergeRequestApi = gitLabMemberUserContext.getGitLabAPI().getMergeRequestApi();
-        Integer parsedMergeRequestId = Integer.parseInt(reviewId);
-        Integer gitlabProjectId = sdlcGitLabProjectId.getGitLabId();
+        Long parsedMergeRequestId = Long.parseLong(reviewId);
+        Long gitlabProjectId = sdlcGitLabProjectId.getGitLabId();
 
         String requiredStatus = "can_be_merged";
         CallUntil<MergeRequest, GitLabApiException> callUntil = CallUntil.callUntil(
